Fix lance incremental upsert, update deps, add 7 language extractors
- persist_graph_incremental: uses merge_insert (upsert+sync) instead of DROP+CREATE,
  eliminating the empty-result window during incremental graph updates
- update_graph_incrementally now calls persist_graph_incremental

Dep updates:
- tree-sitter 0.25→0.26, python 0.23→0.25, go 0.23→0.25, petgraph 0.7→0.8,
  sqlparser 0.59→0.61, pulldown-cmark 0.12→0.13 (arrow stays at 57: lancedb 0.26 constraint)
- Fix breaking changes: tree-sitter child(u32) cast, sqlparser ForeignKey/AlterTable newtype variants

New extractors: JavaScript (js/jsx/mjs), Java, Bash (sh), HCL/Terraform (tf/hcl/tfvars),
JSON, TOML, YAML (with k8s manifest detection)

+++ src/code/mod.rs
@@ -79,7 +79,7 @@ fn collect_symbols(
         if kind == SymbolKind::Impl {
             let scope = Some(name);
             for i in 0..node.child_count() {
-                if let Some(child) = node.child(i) {
+                if let Some(child) = node.child(i as u32) {
                     collect_symbols(child, path, source, &scope, symbols);
                 }
             }
@@ -89,7 +89,7 @@ fn collect_symbols(
 
     // Recurse into children for non-impl nodes (or nodes that aren't symbols)
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i) {
+        if let Some(child) = node.child(i as u32) {
             collect_symbols(child, path, source, parent_scope, symbols);
         }
     }
